requirements.txt
PyPDF2
Pillow
reportlab
Flask-Mail